Extract commit subcommand to its own module
